9-times_table.c

+++ 0x02-functions_nested_loops/9-times_table.c
@@ -7,7 +7,7 @@
  */
 void times_table(void)
 {
-	int double a;
+  double a;
 
 	for (a = 0; a <= 9; a++)
 		{
